requirements.txt
# Core ML/AI
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
openai>=1.0.0
anthropic>=0.3.0

# Multi-Agent Framework
langchain>=0.1.0
langchain-community>=0.1.0
langchain-openai>=0.1.0
crewai>=0.1.0
autogen-agentchat>=0.2.0

# Tools and Integrations
tavily-python>=0.3.0
duckduckgo-search>=4.0.0
wikipedia>=1.4.0
requests>=2.31.0

# Vector Database & Search
chromadb>=0.4.0
faiss-cpu>=1.7.4
rank-bm25>=0.2.2

# Document Processing
pypdf2>=3.0.0
pymupdf>=1.23.0
python-docx>=0.8.11
unstructured>=0.10.0
beautifulsoup4>=4.12.0
markdown>=3.5.0
pyyaml>=6.0.0

# Web Framework & API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Data & Utils
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
nltk>=3.8.0
spacy>=3.6.0

# Async & Concurrency
asyncio
aiofiles>=23.0.0
httpx>=0.24.0

# Monitoring & Logging
structlog>=23.0.0
prometheus-client>=0.17.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
pytest-cov>=4.1.0

# UI Framework
streamlit>=1.28.0
plotly>=5.17.0
streamlit-chat>=0.1.0

# Evaluation & Monitoring
ragas>=0.1.0
langsmith>=0.1.0
wandb>=0.16.0
datasets>=2.14.0

# Memory & Persistence
redis>=5.0.0

# Additional ML/AI
torch>=2.0.0
scikit-learn>=1.3.0